# xhs_backend/api_docs/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "小红书评论维护系统",
    "description": "小红书评论数据的管理和维护系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/comments": {
      "get": {
        "tags": [
          "评论"
        ],
        "summary": "Get Comments",
        "description": "获取评论列表，支持多种过滤条件\n\nArgs:\n    keyword: 关键词（评论内容）\n    status: 评论状态\n    noteId: 笔记ID\n    authorName: 作者名称\n    startDate: 开始日期\n    endDate: 结束日期\n    pagination: 分页参数\n    current_user: 当前用户\n    \nReturns:\n    评论列表和总数",
        "operationId": "get_comments_api_comments_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Keyword"
            },
            "name": "keyword",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            },
            "name": "status",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Noteid"
            },
            "name": "noteId",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorname"
            },
            "name": "authorName",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Startdate"
            },
            "name": "startDate",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Enddate"
            },
            "name": "endDate",
            "in": "query"
          },
          {
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1.0,
              "title": "Page",
              "description": "页码",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "description": "每页项目数",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100.0,
              "minimum": 1.0,
              "title": "Page Size",
              "description": "每页项目数",
              "default": 20
            },
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Comments Api Comments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/comments/{comment_id}": {
      "get": {
        "tags": [
          "评论"
        ],
        "summary": "Get Comment By Id",
        "description": "根据ID获取评论详情\n\nArgs:\n    comment_id: 评论ID\n    current_user: 当前用户\n    \nReturns:\n    评论详情",
        "operationId": "get_comment_by_id_api_comments__comment_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            },
            "name": "comment_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Comment By Id Api Comments  Comment Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "评论"
        ],
        "summary": "Delete Comment",
        "description": "删除评论\n\nArgs:\n    comment_id: 评论ID\n    current_user: 当前用户\n    \nReturns:\n    删除结果",
        "operationId": "delete_comment_api_comments__comment_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            },
            "name": "comment_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Comment Api Comments  Comment Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/comments/{comment_id}/status": {
      "put": {
        "tags": [
          "评论"
        ],
        "summary": "Update Comment Status",
        "description": "更新评论状态\n\nArgs:\n    comment_id: 评论ID\n    status_data: 包含新状态的数据\n    current_user: 当前用户\n    \nReturns:\n    更新结果",
        "operationId": "update_comment_status_api_comments__comment_id__status_put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            },
            "name": "comment_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Status Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Update Comment Status Api Comments  Comment Id  Status Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/comments/batch/status": {
      "put": {
        "tags": [
          "评论"
        ],
        "summary": "Batch Update Status",
        "description": "批量更新评论状态\n\nArgs:\n    data: 包含评论ID列表和新状态的数据\n    current_user: 当前用户\n    \nReturns:\n    更新结果",
        "operationId": "batch_update_status_api_comments_batch_status_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Batch Update Status Api Comments Batch Status Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/comments/batch/delete": {
      "post": {
        "tags": [
          "评论"
        ],
        "summary": "Batch Delete",
        "description": "批量删除评论\n\nArgs:\n    data: 包含评论ID列表的数据\n    current_user: 当前用户\n    \nReturns:\n    删除结果",
        "operationId": "batch_delete_api_comments_batch_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Batch Delete Api Comments Batch Delete Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/comments/user/{user_id}": {
      "get": {
        "tags": [
          "评论"
        ],
        "summary": "Get User Comments",
        "description": "获取指定用户ID的所有历史评论\n\nArgs:\n    user_id: 用户ID\n    current_user: 认证用户名\n    \nReturns:\n    包含用户评论及相关笔记信息的列表，按时间降序排序",
        "operationId": "get_user_comments_api_comments_user__user_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            },
            "name": "user_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Get User Comments Api Comments User  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/comments/data": {
      "post": {
        "tags": [
          "评论",
          "数据接收"
        ],
        "summary": "Receive Comments Data",
        "description": "接收评论数据\n\nArgs:\n    payload: 包含评论数据的有效载荷\n    current_user: 当前用户\n    \nReturns:\n    保存结果",
        "operationId": "receive_comments_data_api_comments_data_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomingPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Receive Comments Data Api Comments Data Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncomingPayload": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "通知",
              "评论",
              "笔记"
            ],
            "title": "Type"
          },
          "data": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "type",
          "data"
        ],
        "title": "IncomingPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}